job: move into model/job
